NO-ISSUE: Reorganize command packages
Currently there is one package for each command. This means a large
number of very small packages, and makes it difficult to share things
between related commands. For example, the `create cluster` and `delete
cluster` command and in different packages and they can't easily share
things. To improve that this patch moves related commands to the same
package, for example, the `create cluster` and `delete cluster` commands
will now be part of the same package.
