--- ztp/internal/cmd/create/create_cmd.go
@@ -1,40 +1,0 @@
-/*
-Copyright 2023 Red Hat Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in
-compliance with the License. You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under the License is
-distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-implied. See the License for the specific language governing permissions and limitations under the
-License.
-*/
-
-package create
-
-import (
-	"github.com/spf13/cobra"
-
-	createclustercmd "github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd/create/cluster"
-	createicspcmd "github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd/create/icsp"
-	createlsocmd "github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd/create/lso"
-	createmetallbcmd "github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd/create/metallb"
-	createuicmd "github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd/create/ui"
-)
-
-// Cobra creates and returns the `create` command.
-func Cobra() *cobra.Command {
-	result := &cobra.Command{
-		Use:   "create",
-		Short: "Creates objects",
-		Args:  cobra.NoArgs,
-	}
-	result.AddCommand(createclustercmd.Cobra())
-	result.AddCommand(createicspcmd.Cobra())
-	result.AddCommand(createlsocmd.Cobra())
-	result.AddCommand(createmetallbcmd.Cobra())
-	result.AddCommand(createuicmd.Cobra())
-	return result
-}
