+++ ztp/main.go
@@ -20,10 +20,7 @@ import (
 	"os"
 
 	"github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal"
-	createcmd "github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd/create"
-	deletecmd "github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd/delete"
-	devcmd "github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd/dev"
-	versioncmd "github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd/version"
+	"github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/cmd"
 	"github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/ztp/internal/exit"
 )
 
@@ -37,10 +34,10 @@ func main() {
 		SetIn(os.Stdin).
 		SetOut(os.Stdout).
 		SetErr(os.Stderr).
-		AddCommand(createcmd.Cobra).
-		AddCommand(devcmd.Cobra).
-		AddCommand(versioncmd.Cobra).
-		AddCommand(deletecmd.Cobra).
+		AddCommand(cmd.Create).
+		AddCommand(cmd.Delete).
+		AddCommand(cmd.Dev).
+		AddCommand(cmd.Version).
 		Build()
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "%s\n", err.Error())
